Add a generic transformer js benchmark runner.
Use webpack to bundle transformers js benchmark suite so that all the
dependencies are included as local files. This also makes it relatively
easy to add additional models/devices.

Before testing you'll need `npm install && npm build` in
`resources/transformers-js` folder.

Notes:
- I'm not very familiar with this code base so I hacked up a little
change to allow additional parameters on test entry html page. That way
we can use the same runner to switch between the various models/devices.
There is probably a better way to do this.

- I didn't commit the code in `dist` since that includes the model. I'm
not sure how we want to handle this? i.e. should we require a build step
to download models or do we want to commit all the code statically?